carritoCompleto

+++ goMovie/src/main/java/com/goMovie/Modelo/Pelicula.java
@@ -27,7 +27,7 @@ public class Pelicula {
 	private String nombre;
 	private String descripcion;
 	private String duracion;
-	private float precio;
+	private double precio;
 	private int stock;
 	private int rebaja;
 	private int descatalogado;
@@ -49,7 +49,7 @@ public class Pelicula {
 		super();
 	}
 
-	public Pelicula(int id_pelicula, String foto, String nombre, String descripción, String duracion, float precio,
+	public Pelicula(int id_pelicula, String foto, String nombre, String descripción, String duracion, double precio,
 			int stock, int rebaja, int descatalogado, Genero genero, int puntuacion) {
 		super();
 		this.id_pelicula = id_pelicula;
@@ -106,11 +106,11 @@ public class Pelicula {
 		this.duracion = duracion;
 	}
 
-	public float getPrecio() {
+	public double getPrecio() {
 		return precio;
 	}
 
-	public void setPrecio(float precio) {
+	public void setPrecio(double precio) {
 		this.precio = precio;
 	}
